python/requirements.txt
Flask==3.0.0
plaid_python==24.0.0
python-dotenv==0.15.0
itsdangerous==2.1.2
werkzeug==3.0.1
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6